stage 1, 2 /3: Design
You got to the most creative part of the project: design! It's time to work out what your web-page should look like. You can make the decisions as you go, draw a tentative plan on a paper, or even use a design app like Figma. Feel free to express your creativity, though there are some restrictions: since projects are about learning, restrictions are necessary for educational and testing purposes.

Don't forget to add a CSS file to the src folder and include a link to it in HTML; otherwise, the added styles will not apply.

First of all, let's design our flashcards. Set the size of divs with questions (notice: the cards are square), add color to the background, round the corners if you want, and position the text inside. Then form a table from the cards. We suggest using the following model of a page and CSS properties grid/flex.

Set the background for the element that contains all the cards and the title. We used the picture of the world map, but you can make a solid background if you prefer.

As a final touch, change the fonts of the title and the text on your flashcards.

Be careful, don't use extra elements without need.
Objectives

1. Make a set of square flash cards.
2. Set the background for the cards and the whole page using the CSS property background-image/background-color.
3. Position the cards using the CSS property grid/flex so they make a 3x3 table.
4. Set the fonts for the text.